Refactor hyperconverged monitoring into separate package
