+++ api/v1beta1/zz_generated.deepcopy.go
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * Copyright 2024 Red Hat, Inc.
+ * Copyright 2025 Red Hat, Inc.
  *
  */
