+++ pkg/monitoring/hyperconverged/metrics/infrastructure_metrics.go
