+++ pkg/monitoring/hyperconverged/metrics/operator_metrics.go
