+++ tools/metricsdocs/metricsdocs.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/machadovilaca/operator-observability/pkg/docs"
 
-	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/monitoring/metrics"
-	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/monitoring/rules"
+	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/monitoring/hyperconverged/metrics"
+	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/monitoring/hyperconverged/rules"
 )
 
 const tpl = `# Hyperconverged Cluster Operator metrics
