+++ tools/prom-metrics-collector/metrics_json_generator.go
@@ -25,8 +25,8 @@ import (
 
 	"github.com/kubevirt/monitoring/pkg/metrics/parser"
 
-	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/monitoring/metrics"
-	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/monitoring/rules"
+	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/monitoring/hyperconverged/metrics"
+	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/monitoring/hyperconverged/rules"
 )
 
 // This should be used only for very rare cases where the naming conventions that are explained in the best practices:
